Add files via upload
The Palmoria Group, a manufacturing company based in the Nigeria, it is embroiled in issues bordering on gender inequality in its 3 regions.

the task here is to use my knowledge of Power BI and Dax to analyze and provide solution to the company issues. kindly find attached the power bi document. click the icon